# api/swagger.yaml
...
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "FC Api"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "agnetiuslee@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/ronaldSengkey/fitClub/api/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /partner/class/{param}/{token}:
    get:
      tags:
      - "member"
      summary: "Add new place off partner"
      operationId: "partnerClass"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        description: "token of partner"
        required: true
        type: "string"
      - name: "param"
        in: "path"
        description: "parameter filter"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/notifSuccess"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "member"
...
  /place/{token}:
    get:
      tags:
      - "coach"
      summary: "Registration access"
      description: ""
      operationId: "placeList"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
        404:
          description: "Data not found"
      x-swagger-router-controller: "class"
  /class/schedule/{token}:
    get:
      tags:
      - "coach"
      summary: "Registration access"
      description: ""
      operationId: "classSchedule"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        required: true
        type: "string"
      - name: "param"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
        404:
          description: "Data not found"
      x-swagger-router-controller: "class"    
  /class/{token}:
    get:
      tags:
      - "coach"
      summary: "Registration access"
      description: ""
      operationId: "classList"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        required: true
        type: "string"
      - name: "byClassId"
        in: "header"
        required: false
        type: "string"
      - name: "byDate"
        in: "header"
        required: false
        type: "string"
      - name: "param"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
        404:
          description: "Data not found"
      x-swagger-router-controller: "class"

    post:
      tags:
      - "coach"
      summary: "Registration access"
      description: ""
      operationId: "createClass"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "data credential guest"
        required: true
        schema:
          $ref: "#/definitions/createClassParam"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
        404:
          description: "Data not found"
      x-swagger-router-controller: "class"
  /class/memberClass/history/{token}:
    get:
      tags:
      - "coach"
      summary: "Registration access"
      description: ""
      operationId: "memberClassHistory"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
        404:
          description: "Data not found"
      x-swagger-router-controller: "class"
  /class/coachClass/history/{token}:
    get:
      tags:
      - "class"
      summary: "Registration access"
      description: ""
      operationId: "coachClassHistory"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
        404:
          description: "Data not found"
      x-swagger-router-controller: "class"
  /class/memberClass/{token}:
    get:
      tags:
      - "coach"
      summary: "Registration access"
      description: ""
      operationId: "memberClass"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
        404:
          description: "Data not found"
      x-swagger-router-controller: "class"
  /class/detail/{token}/{classId}:
      get:
        tags:
        - "coach"
        summary: "Registration access"
        description: ""
        operationId: "classDetail"
        consumes:
        - "multipart/form-data"
        - "application/x-www-form-urlencoded"
        - "application/json"
        produces:
        - "application/xml"
        - "application/json"
        parameters:
        - name: "token"
          in: "path"
          required: true
          type: "string"
        - name: "classId"
          in: "path"
          required: true
          type: "integer"
          format: "int8"
        responses:
          200:
            description: "Success"
          500:
            description: "Internal error"
          401:
            description: "Unauthorize"
          404:
            description: "Data not found"
        x-swagger-router-controller: "class"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  createClassParam:
    type: "object"
    properties:
      className:
        type: "string"
      description:
        type: "string"
...
  notifSuccess:
    type: "object"
    properties:
      message:
        type: "string"
      data:
        type: "string"
    example:
      message: "Success"
      data: "data all customer"
...